requirements.txt
pip~=23.1.2
Django~=5.0.7
asgiref~=3.8.1
sqlparse~=0.5.1
channels~=4.1.0
djangorestframework~=3.15.2
setuptools~=65.5.0
psycopg2-binary~=2.9.6
channels-redis~=4.1.0
daphne~=4.0.0
django-cors-headers~=3.13.0